add func

+++ src/main/java/org/geilove/requestParam/CashParam.java
@@ -4,19 +4,21 @@ public class CashParam {
 	
 	private String proof;
 	
-	private Integer cashid; //应该是Long类型，受限于tweet表
-	
+	private Long cashid; //应该是Long类型，受限于tweet表
+
 	public String getProof() {
 		return proof;
 	}
+
 	public void setProof(String proof) {
 		this.proof = proof;
 	}
-	public Integer getCashid() {
+
+	public Long getCashid() {
 		return cashid;
 	}
-	public void setCashid(Integer cashid) {
+
+	public void setCashid(Long cashid) {
 		this.cashid = cashid;
 	}
-	
 }
